Fix up dependencies.
While auditing the package for release, I found a number of
dependencies that weren't declared, or not declared entirely
properly.  There were also unnecessary dependencies declared.
This one fixes all of the ones I noticed, though it will *not*
tell us if we have missed a dependency that we need.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>